add doc, use button instead switch
