+++ custom_components/tesla_custom/const.py
@@ -16,6 +16,7 @@ PLATFORMS = [
     "binary_sensor",
     "device_tracker",
     "switch",
+    "button",
 ]
 
 ICONS = {
@@ -31,8 +32,8 @@ ICONS = {
     "location tracker": "mdi:crosshairs-gps",
     "charging rate sensor": "mdi:speedometer",
     "sentry mode switch": "mdi:shield-car",
-    "horn switch": "mdi:bullhorn",
-    "flash light switch": "mdi:car-light-high",
+    "horn": "mdi:bullhorn",
+    "flash light": "mdi:car-light-high",
     "solar panel": "mdi:solar-panel",
 }
 AUTH_CALLBACK_PATH = "/auth/tesla/callback"
